Add book form to containers
