+++ src/components/BookForm.js
@@ -105,19 +105,4 @@ BookFormComponent.propTypes = {
   id: PropType.number,
 };
 
-const mapStateToProps = (state) => {
-  const { books } = state.bookReducer;
-  const lastBook = books[books.length - 1];
-  const id = lastBook ? lastBook.id + 1 : 1;
-  return {
-    id,
-    categories: state.filterReducer.categories,
-  };
-};
-
-
-const mapDispatchToProps = (dispatch) => ({
-  addNewBook: (book) => dispatch(addBook(book)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(BookFormComponent);
+export default BookFormComponent;
